+++ src/features/preference/model/useGetUserPreferredArtists.ts
@@ -1,10 +1,11 @@
 import { useQuery } from "@tanstack/react-query";
 import getUserPreferredArtists from "../api/getUserPreferredArtists";
 
-function useGetUserPreferredArtists() {
+function useGetUserPreferredArtists(isLoggedIn: boolean) {
   return useQuery({
     queryKey: ["userPreferredArtists"],
     queryFn: getUserPreferredArtists,
+    enabled: isLoggedIn,
   });
 }
 export default useGetUserPreferredArtists;
