+++ src/features/preference/model/useGetUserPreferredGenres.ts
@@ -1,10 +1,11 @@
 import { useQuery } from "@tanstack/react-query";
-import getUserPreferredGenres from "../api/getUerPreferredGenres";
+import getUserPreferredGenres from "../api/getUserPreferredGenres";
 
-function useGetUserPreferredGenres() {
+function useGetUserPreferredGenres(isLoggedIn: boolean) {
   return useQuery({
     queryKey: ["getUserPreferredGenres"],
     queryFn: getUserPreferredGenres,
+    enabled: isLoggedIn,
   });
 }
